api/requirements-base.txt
django==2.2.14
django-environ==0.4.5
django-filter==2.4.0
django-localized-fields==6.2
django-postgres-extra==2.0.2
django_simple_history==2.12.0
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
mozilla-django-oidc==1.2.4
requests==2.25.1
uwsgi==2.0.19.1
